+++ src/main/java/br/com/moodvie/services/TMDBService.java
@@ -18,11 +18,11 @@ import java.util.List;
 public class TMDBService {
     private final TMDBRepository tmdbRepository;
 
-    public TMDBResponse getRecommendation(String type, Integer contentId,Integer pageNumber,String lang){
+    public TMDBResponse getRecommendation(String type, Long contentId,Integer pageNumber,String lang){
         return tmdbRepository.findTMDBrecommendations(type,contentId,pageNumber,lang);
     }
 
-    public ContentDetailsTMDB getMovieDetails(String type, Integer contentId,String lang){
+    public ContentDetailsTMDB getMovieDetails(String type, Long contentId, String lang){
         Object movieDetailsObject = tmdbRepository.findTMDBDetails(type,contentId,lang);
         ContentDetailsTMDB movieDetails;
         try{
